+++ ThaumAge/Assets/Editor/Game/BlockEditorWindow.cs
@@ -154,7 +154,7 @@ public class BlockEditorWindow : EditorWindow
         {
             if (EditorUI.GUIButton("创建方块", 150))
             {
-                blockInfo.link_id = blockInfo.id;
+                blockInfo.link_id = (int)blockInfo.id;
                 blockInfo.valid = 1;
                 bool isSuccess = serviceForBlockInfo.UpdateData(blockInfo);
                 if (!isSuccess)
@@ -167,7 +167,7 @@ public class BlockEditorWindow : EditorWindow
         {
             if (EditorUI.GUIButton("更新方块", 150))
             {
-                blockInfo.link_id = blockInfo.id;
+                blockInfo.link_id = (int)blockInfo.id;
                 bool isSuccess = serviceForBlockInfo.UpdateData(blockInfo);
                 if (!isSuccess)
                 {
