+++ ThaumAge/Assets/Scrpits/Game/Block/Base/BlockBaseLink.cs
@@ -14,7 +14,7 @@ public class BlockBaseLink : Block
         BlockMetaBaseLink blockMetaBaseLink = new BlockMetaBaseLink();
         blockMetaBaseLink.level = 0;
         blockMetaBaseLink.linkBasePosition = new Vector3IntBean(worldPosition);
-        blockMetaBaseLink.baseBlockType = blockInfo.id;
+        blockMetaBaseLink.baseBlockType = (int)blockInfo.id;
         return ToMetaData(blockMetaBaseLink);
     }
